bug 1441700 - Move JobDetailsPane to details-panel folder
