+++ ui/details-panel/JobDetailsPane.jsx
